numpy Shape Manipulation
the shape of an array is determined by the number of elements on each axis of the array. As such, three numpy functions can be used to change the shape of the array to a way you want: // np.ravel() - this flattens the array using a C-Style method i.e. the rightmost index changes the fastest. // np.T() - this returns an array transposed, i.e. the elements of an array are interchanged in place // np.reshape() - this returns the array with a modified shape.//